Move spaghetti up one level, allcountries, add blank orders pane
