+++ beta-src/src/components/controllers/WDMapController.tsx
@@ -10,20 +10,11 @@ import { useAppDispatch, useAppSelector } from "../../state/hooks";
 import {
   gameApiSliceActions,
   gameOrdersMeta,
-  gameOverview,
-  gameStatus,
-  gameData,
-  gameMaps,
-  gameViewedPhase,
-  gameLegalOrders,
 } from "../../state/game/game-api-slice";
-import {
-  Unit,
-  getUnitsLive,
-  getUnitsHistorical,
-} from "../../utils/map/getUnits";
-import { IOrderData, IOrderDataHistorical } from "../../models/Interfaces";
-import provincesMapData from "../../data/map/ProvincesMapData";
+import { Unit } from "../../utils/map/getUnits";
+import { IOrderDataHistorical } from "../../models/Interfaces";
+import GameStateMaps from "../../state/interfaces/GameStateMaps";
+import { APITerritories } from "../../state/interfaces/GameDataResponse";
 
 const Scales: Scale = {
   DESKTOP: [0.45, 3],
@@ -42,199 +33,31 @@ const getInitialScaleForDevice = (device: Device): number[] => {
 const mapOriginalWidth = 6010;
 const mapOriginalHeight = 3005;
 
-// TODO big spaghetti to unify webdip's historical representation
-// with webdip's live representation
-// and with ordersMeta.
-// into one single set of data about orders and units that we can
-// pass down so that everything else below us renders functionally
-// based on that.
-const WDMapController: React.FC = function (): React.ReactElement {
+interface WDMapControllerProps {
+  units: Unit[];
+  phase: string;
+  orders: IOrderDataHistorical[];
+  maps: GameStateMaps;
+  territories: APITerritories;
+  centersByProvince: { [key: string]: { ownerCountryID: string } };
+  isLatestPhase: boolean;
+}
+
+const WDMapController: React.FC<WDMapControllerProps> = function ({
+  units,
+  phase,
+  orders,
+  maps,
+  territories,
+  centersByProvince,
+  isLatestPhase,
+}): React.ReactElement {
   const svgElement = React.useRef<SVGSVGElement>(null);
   const [viewport] = useViewport();
   const dispatch = useAppDispatch();
   const ordersMeta = useAppSelector(gameOrdersMeta);
   const device = getDevice(viewport);
   const [scaleMin, scaleMax] = getInitialScaleForDevice(device);
-
-  // FIXME: it's not ideal for us to be fetching the whole world from store here
-  // This is hard to untangle though because the representation of the data in the
-  // store is relatively bad. You have to depend on a lot of stuff in order to
-  // draw useful things right now.
-  const viewedPhaseState = useAppSelector(gameViewedPhase);
-  const overview = useAppSelector(gameOverview);
-  const status = useAppSelector(gameStatus);
-  const data = useAppSelector(gameData);
-  const maps = useAppSelector(gameMaps);
-
-  const updateForPhase = () => {
-    if (
-      viewedPhaseState.viewedPhaseIdx >= status.phases.length - 1 &&
-      status.status === "Playing"
-    ) {
-      // Convert from our internal order representation to webdip's
-      // historical representation of orders so that we draw
-      // our internal orders and webdip's historical orders
-      // exactly the same way.
-
-      const ordersHistorical: IOrderDataHistorical[] = [];
-      const currentOrdersById: { [key: number]: IOrderData } = {};
-      if (data.data.currentOrders) {
-        data.data.currentOrders.forEach((orderData) => {
-          currentOrdersById[orderData.id] = orderData;
-        });
-      }
-      Object.entries(ordersMeta).forEach(([orderID, orderMeta]) => {
-        // FIXME ordersMeta can accumulate garbage over multiple phases.
-        // Is there anywhere else where we iterate over it and therefore iterate
-        // over garbage orders?
-        if (!currentOrdersById[orderID]) {
-          return;
-        }
-        let fromTerrID = 0;
-        let toTerrID = 0;
-        let terrID = 0;
-        let type: string | null = "";
-        let unitType = "";
-        let viaConvoy;
-
-        let { originalOrder } = orderMeta;
-        if (!originalOrder) {
-          originalOrder = currentOrdersById[orderID];
-        }
-        if (originalOrder) {
-          if (originalOrder.fromTerrID) {
-            fromTerrID = Number(originalOrder.fromTerrID);
-          }
-          if (originalOrder.toTerrID) {
-            toTerrID = Number(originalOrder.toTerrID);
-          }
-          type = originalOrder.type;
-
-          if (type && type.startsWith("Build ")) {
-            if (originalOrder.toTerrID) {
-              terrID = Number(originalOrder.toTerrID);
-            }
-            [, unitType] = type.split(" ");
-          } else if (originalOrder.unitID) {
-            const terrIDString = maps.unitToTerrID[originalOrder.unitID];
-            if (terrIDString) {
-              terrID = Number(terrIDString);
-            }
-            unitType = data.data.units[originalOrder.unitID].type;
-          }
-
-          if (originalOrder.viaConvoy === "Yes") {
-            viaConvoy = "Yes";
-          } else {
-            viaConvoy = "No";
-          }
-        }
-        if (orderMeta.update) {
-          if (orderMeta.update.fromTerrID !== undefined) {
-            fromTerrID = Number(orderMeta.update.fromTerrID);
-          }
-          toTerrID = Number(orderMeta.update.toTerrID);
-          type = orderMeta.update.type;
-          if (orderMeta.update.viaConvoy === "Yes") {
-            viaConvoy = "Yes";
-          } else {
-            viaConvoy = "No";
-          }
-        }
-
-        // !terrID is safe because webdip doesn't seem to use terrID 0.
-        if (!type || !unitType || !terrID) {
-          return;
-        }
-
-        const orderHistorical: IOrderDataHistorical = {
-          countryID: status.countryID,
-          dislodged: "No",
-          fromTerrID,
-          phase: overview.phase,
-          success: "Yes",
-          terrID,
-          toTerrID,
-          turn: overview.turn,
-          type,
-          unitType,
-          viaConvoy,
-        };
-        ordersHistorical.push(orderHistorical);
-      });
-      // console.log("Ordershistorical");
-      // console.log(currentOrdersById);
-      // console.log(state.game.ordersMeta);
-      // console.log(ordersHistorical);
-
-      // Also depends on status, so this is updated both here and when GameStatus is fulfilled.
-      const prevPhaseOrders =
-        status.phases.length > 1
-          ? status.phases[status.phases.length - 2].orders
-          : [];
-      const units: Unit[] = getUnitsLive(
-        data.data.territories,
-        data.data.territoryStatuses,
-        data.data.units,
-        overview.members,
-        prevPhaseOrders,
-        ordersMeta,
-        data.data.currentOrders ? data.data.currentOrders : [],
-        overview.user,
-        overview.phase,
-        maps,
-      );
-
-      const centersByProvince: { [key: string]: { ownerCountryID: string } } =
-        {};
-      data.data.territoryStatuses.forEach((provinceStatus) => {
-        const province = maps.terrIDToProvince[provinceStatus.id];
-        const ownerCountryID = provinceStatus.ownerCountryID || "0";
-        centersByProvince[province] = { ownerCountryID };
-      });
-
-      return {
-        phase: overview.phase,
-        units,
-        orders: ordersHistorical,
-        centersByProvince,
-        isLatestPhase: true,
-      };
-    }
-
-    const phaseHistorical = status.phases[viewedPhaseState.viewedPhaseIdx];
-    const unitsHistorical = phaseHistorical.units;
-    const prevPhaseOrders =
-      viewedPhaseState.viewedPhaseIdx > 0
-        ? status.phases[viewedPhaseState.viewedPhaseIdx - 1].orders
-        : [];
-    const unitsLive = getUnitsHistorical(
-      data.data.territories,
-      unitsHistorical,
-      overview.members,
-      prevPhaseOrders,
-      phaseHistorical.orders,
-      maps,
-    );
-    const centersByProvince: {
-      [key: string]: { ownerCountryID: string };
-    } = {};
-    phaseHistorical.centers.forEach((iCenter) => {
-      centersByProvince[maps.terrIDToProvince[iCenter.terrID]] = {
-        ownerCountryID: iCenter.countryID.toString(),
-      };
-    });
-    return {
-      phase: phaseHistorical.phase as string,
-      units: unitsLive,
-      orders: phaseHistorical.orders,
-      centersByProvince,
-      isLatestPhase: false,
-    };
-  };
-  const { phase, units, orders, centersByProvince, isLatestPhase } =
-    updateForPhase();
-  const { territories } = data.data;
 
   // const legalOrders = useAppSelector(gameLegalOrders);
   // console.log({ legalOrders });
